+++ src/edu4.Application/Contracts/IAccountManagementService.cs
@@ -1,4 +1,4 @@
-namespace edu4.Application.External;
+namespace edu4.Application.Contracts;
 public interface IAccountManagementService
 {
     public Task MarkUserSignedUpAsync(string accountId);
