+++ src/edu4.Application/Contracts/IUsersRepository.cs
@@ -1,11 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using edu4.Domain.Users;
 
-namespace edu4.Application;
+namespace edu4.Application.Contracts;
 public interface IUsersRepository
 {
     public Task AddAsync(User user);
