+++ src/edu4.Application/UsersService.cs
@@ -1,4 +1,4 @@
-using edu4.Application.External;
+using edu4.Application.Contracts;
 using edu4.Domain.Users;
 using Microsoft.Extensions.Logging;
